Client must explicitely ask for direct blob data access
For compatibility, existing client will get a modifiable copy of data
